不具合を修正

+++ Project-HEW2020/debug_font.cpp
@@ -28,7 +28,7 @@ static Sprite* g_pSpriteDebug = NULL;
 void DebugFont_Init(void)
 {
     g_pSpriteDebug = new SpriteNormal(TEXTURE_DEBUG);
-    g_pSpriteDebug->SetCutRange((float)DEBUG_FONT_WIDTH, (float)DEBUG_FONT_HEIGHT);
+    g_pSpriteDebug->SetCutRange(DEBUG_FONT_WIDTH, DEBUG_FONT_HEIGHT);
     g_pSpriteDebug->SetPolygonSize(DEBUG_FONT_DRAW_WIDTH, DEBUG_FONT_DRAW_HEIGHT);
     g_pSpriteDebug->SetColor(D3DCOLOR_RGBA(255, 200, 150, 255));
 }
@@ -46,8 +46,8 @@ void DebugFont_Draw(float dx, float dy, const char *pString)
     for (int i = 0; i < strlen(pString); i++)
     {
         int offset = pString[i] - ' ';
-        float tcx = (float)(offset % DEBUG_FONT_START_CHAR_CODE * DEBUG_FONT_WIDTH);
-        float tcy = (float)(offset / DEBUG_FONT_ONE_LINE_CHAR_COUNT * DEBUG_FONT_HEIGHT);
+        int tcx = (offset % DEBUG_FONT_START_CHAR_CODE * DEBUG_FONT_WIDTH);
+        int tcy = (offset / DEBUG_FONT_ONE_LINE_CHAR_COUNT * DEBUG_FONT_HEIGHT);
 
         g_pSpriteDebug->SetCutPos(tcx, tcy);
         g_pSpriteDebug->SetDrawPos(dx + i * DEBUG_FONT_DRAW_WIDTH, dy);
